fixed can message memory accesses for reading and sending to tritium
fixed cruise velocity setpoint to work with rpm instead of m/s

+++ Apps/Inc/ReadTritium.h
@@ -34,8 +34,8 @@ typedef enum{
  */
 tritium_error_code_t MotorController_getTritiumError(void);
 
-uint32_t Motor_RPM_Get();
-uint32_t Motor_Velocity_Get();
+float Motor_RPM_Get();
+float Motor_Velocity_Get();
 void MotorController_Restart();
 
 #endif
